Handle new API for zmsg_encode()/zmsg_decode()
The API for zmsq_encode() and zmsg_decode() changed in CZMQ.  Newer
versions of CZMQ encode/decode to/from frames rather than buffers.
To allow rbczmq to work with both old and new CZMQ versions, the rbczmq
API has changed slightly.

The Message.encode method now always encodes to a Frame even if the
underlying CZMQ's zmsg_encode() function encodes to a buffer.  This
should have minimal impact on existing code since Frame has a to_s
method that returns the Frame data as a String.

The Message.decode(arg) method now allows arg to be a String or a Frame
regardless of the underlying CZMQ's zmsg_decode() implementation.  This
provides maximum flexibility and compatibility for users.

This commit has only been tested with the newer CZMQ encode/decode API,
so there may be latent bugs in the implementation for the older CZMQ
API.